refactor: refactoring of videos controllers

+++ src/controllers/videos/delete.js
@@ -1,11 +1,22 @@
 const prisma = require("../../database/prisma");
 const errorMessages = require("../../helpers/codeMessages/errorMessages");
 const sucessMessagesRoadmap = require("../../helpers/codeMessages/roadmapSucessMessages");
+const { getVideo } = require("../../models/Video");
 
 const deleteVideo = async (req, res) => {
   const { id } = req.body;
 
   try {
+    const findVideo = await getVideo(id);
+
+    if (!findVideo) {
+      return res.status(404).json({
+        statusCode: 404,
+        message: errorMessages.errorProcessingThisRequest,
+        body: {},
+      });
+    }
+
     const exclude = await prisma.$transaction(async () => {
       const exclude = await prisma.videosRoadmap.deleteMany({
         where: {
